requirements.txt
Flask==0.12.2
gunicorn==19.6.0
requests==2.25.1
boto3==1.18.47
pymessager==1.0.1
google-cloud-vision==2.4.2